modal markup placed in separated file, set LS option added to 'add to watched' btn
